Refactor AI prompt rendering with improved modularity and error handling
